Configuration file appended - https://github.com/purnima04/Hospital_Management_REST/issues/7#issue-910730543

+++ src/main/java/com/banfico/hospital_management/Dao/BillingRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BillingRepository extends JpaRepository<BillingEntity, Long> {
+public interface BillingRepository extends JpaRepository<BillingEntity, Integer> {
 }
